refactor snippet endpoint
